crdbtest: move cockroach colblk.KeySchema
Move the test implementation of a cockroach colblk.KeySchema into crdbtest so
that it may be used by other packages, and most notably, the ycsb benchmarks
implemented in cmd/pebble.
